split code into smaller functions and seperate files to make it more readable
