+++ src/pages/singleEvent/singleEvent.html
@@ -127,5 +127,6 @@
     ></script>
 
     <script type="module" src="/src/pages/singleEvent/ts/posts.ts"></script>
+    <script type="module" src="/src/pages/singleEvent/ts/comment.ts"></script>
   </body>
 </html>
